+++ src/enums/usage-order-by.ts
@@ -1,0 +1,4 @@
+export enum UsageOrderBy {
+    Time = 'time',
+    Value = 'value',
+}